refactor(host-contracts): rename DAO upgrade machinery to upgrade proposals
daoUpgrade named the executor, not the artifact; the refactor's point is that
the executor varies (DAO signs the proposal, devnet executes it directly).
tasks/utils/upgradeProposal.ts now exports UpgradeProposal,
buildUpgradeProposal, printUpgradeProposal, executeUpgradeProposal, and
verifyProposalImplementation; buildCanonicalUpgradeProposal follows suit.
Scoped to this PR's files; wider adoption deferred.
