+++ host-contracts/tasks/taskDeploy.ts
@@ -7,15 +7,15 @@ import type { HardhatEthersHelpers, HardhatRuntimeEnvironment, TaskArguments } f
 import path from 'path';
 
 import {
+  buildCanonicalUpgradeProposal,
   buildSnapshotArtifact,
   parseSnapshotArtifact,
-  prepareCanonicalSnapshotUpgrade,
   readCanonicalSnapshot,
 } from './protocolConfigMirror';
-import { executePreparedDaoUpgrade } from './utils/daoUpgrade';
 import { formatError } from './utils/formatError';
 import { CRS_COUNTER_BASE, KEY_COUNTER_BASE } from './utils/kmsGenerationConstants';
 import { getRequiredEnvVar } from './utils/loadVariables';
+import { executeUpgradeProposal } from './utils/upgradeProposal';
 
 const ADDRESSES_DIR = path.join(__dirname, '../addresses');
 const HOST_ENV_FILE = path.join(ADDRESSES_DIR, '.env.host');
@@ -598,8 +598,8 @@ task(
     // Same prepare step as the DAO path (task:prepareDeployProtocolConfigFromCanonical), then
     // execute the produced payload directly: devnet runs byte-identical calldata to what the DAO
     // would sign.
-    const prepared = await prepareCanonicalSnapshotUpgrade(hre, { snapshot, proxyAddress: secondaryProxyAddress });
-    await executePreparedDaoUpgrade(hre, prepared);
+    const prepared = await buildCanonicalUpgradeProposal(hre, { snapshot, proxyAddress: secondaryProxyAddress });
+    await executeUpgradeProposal(hre, prepared);
 
     // On interval-mining networks, upgradeProxy can return before the tx is mined.
     await waitForTaskReady(hre, 'task:assertProtocolConfigReady');
